Moved authenticate to its own module to avoid circular imports
